# frontend/web/src/api/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LinkRequest": {
    "properties": {
     "email": {
      "maxLength": 254,
      "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
      "title": "Email",
      "type": "string"
     },
     "lang": {
      "default": "zh-TW",
      "pattern": "^[a-z]{2}(-[A-Z][A-Za-z]{1,3})?$",
      "title": "Lang",
      "type": "string"
     },
     "next_path": {
      "anyOf": [
       {
        "maxLength": 200,
        "pattern": "^/[^\\s]*$",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Path"
     }
    },
    "required": [
     "email"
    ],
    "title": "LinkRequest",
    "type": "object"
   },
   "Me": {
    "properties": {
     "email": {
      "title": "Email",
      "type": "string"
     },
     "member_until": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Member Until"
     },
     "reader_id": {
      "format": "uuid",
      "title": "Reader Id",
      "type": "string"
     }
    },
    "required": [
     "reader_id",
     "email"
    ],
    "title": "Me",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
   "Verify": {
    "properties": {
     "token": {
      "maxLength": 200,
      "minLength": 10,
      "title": "Token",
      "type": "string"
     }
    },
    "required": [
     "token"
    ],
    "title": "Verify",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "Autora API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/auth/link": {
   "post": {
    "description": "Email a one-time link. Answers 202 whether or not the address has been seen before.",
    "operationId": "send_link_api_auth_link_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LinkRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Send Link",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/auth/logout": {
   "post": {
    "operationId": "logout_api_auth_logout_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "autora_reader",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Reader"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Logout",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/auth/me": {
   "get": {
    "operationId": "me_api_auth_me_get",
    "parameters": [
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "cookie",
      "name": "autora_reader",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Reader"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/Me"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Me Api Auth Me Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Me",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/auth/verify": {
   "post": {
    "operationId": "verify_api_auth_verify_post",
    "parameters": [
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Verify"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Me"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Verify",
    "tags": [
     "auth"
    ]
   }
  },
...
 }
}